move in election data
